index.tsに関数を切り離す
